feat: update lb apis
