+++ bcs-app/backend/apps/network/urls.py
@@ -15,6 +15,7 @@ from django.conf.urls import url
 from .views import service
 from .views.lb import k8s, mesos
 from .views.ingress import mesos as mesos_ingress
+from .views.charts import versions
 
 MESOS_CLUSTER_ID_REGEX = "BCS-((?!K8S)\w)+-[0-9]{5,7}"
 
@@ -53,7 +54,7 @@ urlpatterns = [
     url(r'^api/network/(?P<project_id>\w{32})/k8s/lb/$',
         k8s.NginxIngressListCreateViewSet.as_view({'get': 'list', 'post': 'create'})),
     url(r'^api/network/(?P<project_id>\w{32})/clusters/(?P<cluster_id>[\w.\-]+)/k8s/lb/namespaces/$',
-        k8s.NginxIngressListNamespceViewSet.as_view({'get': 'list'})),
+        k8s.NginxIngressListNamespaceViewSet.as_view({'get': 'list'})),
     url(r'^api/network/(?P<project_id>\w{32})/k8s/lb/(?P<pk>\d+)/$',
         k8s.NginxIngressRetrieveUpdateViewSet.as_view({'get': 'retrieve', 'put': 'update', 'delete': 'destroy'})),  # noqa
 
@@ -65,5 +66,7 @@ urlpatterns = [
         mesos_ingress.IngressRetrieveOperateViewSet.as_view({'get': 'retrieve', 'delete': 'delete', 'put': 'update'})),
 
     url(r'^api/k8s_lb/projects/(?P<project_id>\w{32})/chart/versions/$',
-        k8s.IngressControllerViewSet.as_view({"get": "list", "post": "query_chart_version"}))
+        versions.K8SIngressControllerViewSet.as_view({"get": "get_chart_versions"})),
+    url(r'^api/k8s_lb/projects/(?P<project_id>\w{32})/chart/versions/-/detail/$',
+        versions.K8SIngressControllerViewSet.as_view({"post": "get_version_detail"}))
 ]
